# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/admin/products": {
      "post": {
        "operationId": "AdminProductsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Product created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a product (admin only)",
        "tags": [
          "Admin"
        ]
      }
    },
    "/api/v1/admin/products/{id}": {
      "patch": {
        "operationId": "AdminProductsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "minimum": 1,
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Product updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Product not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a product (admin only)",
        "tags": [
          "Admin"
        ]
      },
      "delete": {
        "operationId": "AdminProductsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "minimum": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Product deactivated"
          },
          "404": {
            "description": "Product not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Deactivate a product (soft delete, admin only)",
        "tags": [
          "Admin"
        ]
      }
    },
    "/api/v1/admin/products/{id}/availability": {
      "patch": {
        "operationId": "AdminProductsController_updateAvailability",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "minimum": 1,
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAvailabilityDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Availability updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Toggle whether a product can currently be ordered (admin only)",
        "tags": [
          "Admin"
        ]
      }
    },
    "/api/v1/admin/products/{id}/image": {
      "post": {
        "operationId": "AdminProductsController_uploadImage",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "minimum": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Image attached",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Upload or replace the product image (admin only)",
        "tags": [
          "Admin"
        ]
      },
      "delete": {
        "operationId": "AdminProductsController_removeImage",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "minimum": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Image removed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Remove the product image (admin only)",
        "tags": [
          "Admin"
        ]
      }
    },
    "/api/v1/admin/orders": {
      "get": {
        "operationId": "AdminOrdersController_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number (default 1)",
            "schema": {
              "minimum": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Page size (1-100, default 20)",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "Pagination offset (default 0)",
            "schema": {
              "minimum": 0,
              "example": 0,
              "type": "number"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "PENDING",
                "CONFIRMED",
                "PREPARING",
                "READY",
                "OUT_FOR_DELIVERY",
                "COMPLETED",
                "CANCELLED",
                "REJECTED"
              ]
            }
          },
          {
            "name": "orderType",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "DELIVERY",
                "PICKUP"
              ]
            }
          },
          {
            "name": "orderNumber",
            "required": false,
            "in": "query",
            "schema": {
              "example": "BW-20260713-0042",
              "type": "string"
            }
          },
          {
            "name": "customerPhone",
            "required": false,
            "in": "query",
            "schema": {
              "example": "+22200000000",
              "type": "string"
            }
          },
          {
            "name": "fromDate",
            "required": false,
            "in": "query",
            "schema": {
              "example": "2026-07-01",
              "type": "string"
            }
          },
          {
            "name": "toDate",
            "required": false,
            "in": "query",
            "schema": {
              "example": "2026-07-31",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated order list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderListResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List orders with filters (admin only)",
        "tags": [
          "Admin"
        ]
      }
    },
    "/api/v1/admin/orders/{orderNumber}": {
      "get": {
        "operationId": "AdminOrdersController_getOrder",
        "parameters": [
          {
            "name": "orderNumber",
            "required": true,
            "in": "path",
            "schema": {
              "minLength": 1,
              "maxLength": 50,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Order details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Order not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get order details, including invoice information (admin only)",
        "tags": [
          "Admin"
        ]
      }
    },
    "/api/v1/admin/orders/{orderNumber}/status": {
      "patch": {
        "operationId": "AdminOrdersController_updateStatus",
        "parameters": [
          {
            "name": "orderNumber",
            "required": true,
            "in": "path",
            "schema": {
              "minLength": 1,
              "maxLength": 50,
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOrderStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Order status updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid transition",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Transition an order to a new status (admin only)",
        "tags": [
          "Admin"
        ]
      }
    },
...
    "/api/v1/orders/{orderNumber}": {
      "get": {
        "operationId": "OrdersController_getOrder",
        "parameters": [
          {
            "name": "orderNumber",
            "required": true,
            "in": "path",
            "schema": {
              "minLength": 1,
              "maxLength": 50,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Order details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Order not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Get an order by number (owner, or guest with X-Order-Token)",
        "tags": [
          "Orders"
        ]
      }
    },
...
    "/api/v1/me/orders/{orderNumber}": {
      "get": {
        "operationId": "MyOrdersController_getMyOrder",
        "parameters": [
          {
            "name": "orderNumber",
            "required": true,
            "in": "path",
            "schema": {
              "minLength": 1,
              "maxLength": 50,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Order details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Order not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get one of the current user's orders",
        "tags": [
          "Orders"
        ]
      }
    },
...
  },
  "info": {
    "title": "BW Café Ordering API",
    "description": "Single-restaurant delivery/pickup ordering API — guest and registered checkout, admin order management.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ErrorDetailDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "number",
            "example": 409
          },
          "message": {
            "type": "string",
            "example": "Conflict"
          },
          "timestamp": {
            "type": "string",
            "example": "2026-03-28T12:00:00.000Z"
          },
          "path": {
            "type": "string",
            "example": "/api/auth/register/request-otp"
          }
        },
        "required": [
          "code",
          "message",
          "timestamp",
          "path"
        ]
      },
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "statusCode": {
            "type": "number",
            "example": 409
          },
          "data": {
            "type": "object",
            "nullable": true,
            "additionalProperties": false,
            "example": null
          },
          "error": {
            "$ref": "#/components/schemas/ErrorDetailDto"
          }
        },
        "required": [
          "success",
          "statusCode",
          "data",
          "error"
        ]
      },
...
      "ProductCategory": {
        "type": "string",
        "enum": [
          "coffee",
          "breakfast",
          "burger",
          "shawarma",
          "tacos",
          "drinks"
        ]
      },
      "ProductDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "category": {
            "example": "coffee",
            "allOf": [
              {
                "$ref": "#/components/schemas/ProductCategory"
              }
            ]
          },
          "title": {
            "type": "string",
            "example": "Cappuccino"
          },
          "description": {
            "type": "string",
            "nullable": true,
            "example": "Espresso with steamed milk foam"
          },
          "details": {
            "type": "string",
            "nullable": true,
            "example": "Contains dairy"
          },
          "price": {
            "type": "number",
            "example": 250
          },
          "quantity": {
            "type": "number",
            "nullable": true,
            "example": 50,
            "description": "null means stock is not tracked"
          },
          "imageUrl": {
            "type": "string",
            "nullable": true,
            "example": "https://res.cloudinary.com/demo/image/upload/bw-cafe/products/1.jpg"
          },
          "isAvailable": {
            "type": "boolean",
            "example": true
          },
          "isActive": {
            "type": "boolean",
            "example": true
          },
          "createdAt": {
            "type": "string",
            "example": "2026-03-28T12:00:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "example": "2026-03-28T12:00:00.000Z"
          }
        },
        "required": [
          "id",
          "category",
          "title",
          "price",
          "isAvailable",
          "isActive",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "ProductDataDto": {
        "type": "object",
        "properties": {
          "product": {
            "$ref": "#/components/schemas/ProductDto"
          }
        },
        "required": [
          "product"
        ]
      },
      "ProductResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "statusCode": {
            "type": "number",
            "example": 200
          },
          "data": {
            "$ref": "#/components/schemas/ProductDataDto"
          }
        },
        "required": [
          "success",
          "statusCode",
          "data"
        ]
      },
...
      "CreateProductDto": {
        "type": "object",
        "properties": {
          "category": {
            "description": "Product menu category",
            "example": "coffee",
            "allOf": [
              {
                "$ref": "#/components/schemas/ProductCategory"
              }
            ]
          },
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 255,
            "description": "Product title (1–255 chars)",
            "example": "Cappuccino"
          },
          "description": {
            "type": "string",
            "minLength": 1,
            "maxLength": 2000,
            "description": "Short description (1–2000 chars)",
            "example": "Espresso with steamed milk foam"
          },
          "details": {
            "type": "string",
            "minLength": 1,
            "maxLength": 2000,
            "description": "Additional free-text details",
            "example": "Contains dairy"
          },
          "price": {
            "type": "number",
            "minimum": 0,
            "description": "Price in the local currency",
            "example": 250
          },
          "quantity": {
            "type": "number",
            "minimum": 0,
            "description": "Stock quantity; omit to not track stock",
            "example": 50
          }
        },
        "required": [
          "category",
          "title",
          "price"
        ]
      },
      "UpdateProductDto": {
        "type": "object",
        "properties": {
          "category": {
            "description": "Product menu category",
            "example": "coffee",
            "allOf": [
              {
                "$ref": "#/components/schemas/ProductCategory"
              }
            ]
          },
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 255,
            "description": "Product title (1–255 chars)",
            "example": "Cappuccino"
          },
          "description": {
            "type": "string",
            "minLength": 1,
            "maxLength": 2000,
            "description": "Short description (1–2000 chars)",
            "example": "Espresso with steamed milk foam"
          },
          "details": {
            "type": "string",
            "minLength": 1,
            "maxLength": 2000,
            "description": "Additional free-text details",
            "example": "Contains dairy"
          },
          "price": {
            "type": "number",
            "minimum": 0,
            "description": "Price in the local currency",
            "example": 250
          },
          "quantity": {
            "type": "number",
            "minimum": 0,
            "description": "Stock quantity; omit to not track stock",
            "example": 50
          }
        }
      },
      "UpdateAvailabilityDto": {
        "type": "object",
        "properties": {
          "isAvailable": {
            "type": "boolean",
            "description": "Whether the product can currently be ordered",
            "example": false
          }
        },
        "required": [
          "isAvailable"
        ]
      },
      "OrderItemDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "productId": {
            "type": "number",
            "nullable": true,
            "example": 3
          },
          "productTitle": {
            "type": "string",
            "example": "Cappuccino"
          },
          "unitPrice": {
            "type": "number",
            "example": 250
          },
          "quantity": {
            "type": "number",
            "example": 2
          },
          "lineTotal": {
            "type": "number",
            "example": 500
          }
        },
        "required": [
          "id",
          "productTitle",
          "unitPrice",
          "quantity",
          "lineTotal"
        ]
      },
      "OrderDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "orderNumber": {
            "type": "string",
            "example": "BW-20260713-0042"
          },
          "customerName": {
            "type": "string",
            "example": "Customer Name"
          },
          "customerPhone": {
            "type": "string",
            "example": "+22200000000"
          },
          "orderType": {
            "type": "string",
            "enum": [
              "DELIVERY",
              "PICKUP"
            ],
            "example": "DELIVERY"
          },
          "address": {
            "type": "string",
            "nullable": true,
            "example": "Full delivery address as one string"
          },
          "pickupTime": {
            "type": "string",
            "nullable": true,
            "example": null
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "CONFIRMED",
              "PREPARING",
              "READY",
              "OUT_FOR_DELIVERY",
              "COMPLETED",
              "CANCELLED",
              "REJECTED"
            ],
            "example": "PENDING"
          },
          "screenshotUrl": {
            "type": "string",
            "nullable": true
          },
          "subtotal": {
            "type": "number",
            "example": 500
          },
          "deliveryFee": {
            "type": "number",
            "example": 0
          },
          "total": {
            "type": "number",
            "example": 500
          },
          "currency": {
            "type": "string",
            "example": "MRU"
          },
          "customerNotes": {
            "type": "string",
            "nullable": true
          },
          "cancellationReason": {
            "type": "string",
            "nullable": true
          },
          "rejectionReason": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "example": "2026-07-13T12:00:00.000Z"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderItemDto"
            }
          },
          "guestAccessToken": {
            "type": "string",
            "description": "Only present once, at creation, for guest orders"
          }
        },
        "required": [
          "id",
          "orderNumber",
          "customerPhone",
          "orderType",
          "status",
          "subtotal",
          "deliveryFee",
          "total",
          "currency",
          "createdAt",
          "items"
        ]
      },
      "OrderListDataDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderDto"
            }
          },
          "total": {
            "type": "number",
            "example": 42
          }
        },
        "required": [
          "items",
          "total"
        ]
      },
      "OrderListResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "statusCode": {
            "type": "number",
            "example": 200
          },
          "data": {
            "$ref": "#/components/schemas/OrderListDataDto"
          }
        },
        "required": [
          "success",
          "statusCode",
          "data"
        ]
      },
      "OrderDataDto": {
        "type": "object",
        "properties": {
          "order": {
            "$ref": "#/components/schemas/OrderDto"
          }
        },
        "required": [
          "order"
        ]
      },
      "OrderResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "statusCode": {
            "type": "number",
            "example": 200
          },
          "data": {
            "$ref": "#/components/schemas/OrderDataDto"
          }
        },
        "required": [
          "success",
          "statusCode",
          "data"
        ]
      },
      "UpdateOrderStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "CONFIRMED",
              "PREPARING",
              "READY",
              "OUT_FOR_DELIVERY",
              "COMPLETED",
              "CANCELLED",
              "REJECTED"
            ],
            "example": "CONFIRMED"
          },
          "reason": {
            "type": "string",
            "minLength": 1,
            "maxLength": 500,
            "description": "Required when rejecting or cancelling",
            "example": "Out of stock"
          },
          "note": {
            "type": "string",
            "minLength": 1,
            "maxLength": 500,
            "example": "Confirmed after reviewing screenshot"
          }
        },
        "required": [
          "status"
        ]
      },
...
    }
  }
}